FIX: Valgrind leaks

+++ src/config.hpp
@@ -4,7 +4,7 @@
 namespace config {
 namespace geometry {
 /** accuracy of the geometry representation, choose float or double */
-using precision_t = double;
+using precision_t = float;
 } // namespace geometry
 namespace scenario {
